refactor(admin): Q9.c — extract admin/mcp.rs (GAR-471)
Move the four MCP server CRUD handlers out of admin/handlers.rs (2900 LOC)
into a new focused module admin/mcp.rs, following the pattern established by
Q9.a (shared.rs) and Q9.b (providers.rs).

Handlers moved (zero behavior change):
  - admin_list_mcp   (GET /admin/api/mcp)
  - CreateMcpRequest (request body type)
  - admin_create_mcp (POST /admin/api/mcp)
  - admin_restart_mcp (POST /admin/api/mcp/{id}/restart)
  - admin_delete_mcp  (DELETE /admin/api/mcp/{id})

handlers.rs: 2900 → 2550 LOC (-350 LOC)
mcp.rs: new file, 366 LOC

Re-exports in handlers.rs keep routes.rs paths
(handlers::admin_list_mcp etc.) resolving without change.

MCP Templates (list_mcp_templates, save_mcp_template, delete_mcp_template)
are NOT moved — those belong to Q9.d (GAR-472 / plan 0130).
